Fixed compiler warning; Added FIXME.
git-svn-id: file:///home/survex-svn/survex/trunk@1385 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/datain.c
@@ -83,6 +83,8 @@ push_back(int c)
 static void
 error_list_parent_files(void)
 {
+   /* FIXME: don't list this a second time if we just reported another error
+    * in the same file included in the same way... */
    if (file.parent) {
       parse *p = file.parent;
       const char *m = msg(/*In file included from */5);
@@ -92,7 +94,7 @@ error_list_parent_files(void)
       while (p) {
 	 fprintf(STDERR, "%s:%d", p->filename, p->line);
 	 p = p->parent;
-	 if (p) fprintf(STDERR, ",\n%*s", len, m);
+	 if (p) fprintf(STDERR, ",\n%*s", (int)len, m);
       }
       fprintf(STDERR, ":\n");
    }
